--- a/lib/reqIdGenFactory.js
+++ b/lib/reqIdGenFactory.js
@@ -9,6 +9,7 @@ module.exports = function () {
   const maxInt = 2147483647
   let nextReqId = 0
   return function genReqId (req) {
-    return (nextReqId = (nextReqId + 1) & maxInt)
+    nextReqId = (nextReqId + 1) & maxInt
+    return `req-${nextReqId.toString(36)}`
   }
 }