@mcollina 2021-02-12
This needs to be benchmarked before/after.