@mcollina 2021-02-12
try running `npm run benchmark` before/after