@ognjenjevremovic 2021-02-12
Here are the outputs of 3 consecutive runs on before and after the commit.

## Before: 

### Latency

| #Run | 2.5% | 50% | 97.5% | 99% | Avg | Stdev | Max |
| :--- | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
| 1 | 15 ms | 34 ms  | 139 ms | 153 ms | 38.35 ms | 24.69 ms | 298 ms |
| 2 | 18 ms | 46 ms | 181 ms | 285 ms | 59.38 ms | 46.84 ms | 379 ms |
| 3 | 15 ms | 34 ms | 101 ms | 195 ms | 38.77 ms | 32.03 ms | 549 ms |

### Req / Sec

| #Run | 1% | 2.5% | 50% | 97.5% | Avg | Stdev | Min |
| :--- | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
| 1 | 11975 | 11975 | 30415 | 30559 | 25815.2 | 7165.46 | 11968 |
| 2 | 9031 | 9031 | 15871 | 24895 | 16778.41 | 5806.64 | 9030 |
| 3 | 10343 | 10343 | 28575 | 31999 | 25476 | 7879.35 | 10342 |

### Bytes / Sec
| #Run | 1% | 2.5% | 50% | 97.5% | Avg | Stdev | Min |
| :--- | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
| 1 | 2.24 MB | 2.24 MB | 5.69 MB | 5.71 MB | 4.83 MB | 1.34 MB | 2.24 MB |
| 2 | 1.69 MB | 1.69 MB | 2.97 MB | 4.65 MB | 3.14 MB | 1.08 MB | 1.69 MB |
| 3 | 1.93 MB | 1.93 MB | 5.34 MB | 5.98 MB | 4.76 MB | 1.47 MB | 1.93 MB |

---

## After: 

### Latency


| #Run | 2.5% | 50% | 97.5% | 99% | Avg | Stdev | Max |
| :--- | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
| 1 | 16 ms | 37 ms | 129 ms | 168 ms | 45 ms | 32.56 ms | 447 ms |
| 2 | 16 ms | 35 ms | 106 ms | 172 ms | 38.48 ms | 25.11 ms | 256 ms |
| 3 | 15 ms | 33 ms | 91 ms | 154 ms | 38.93 ms | 26.16 ms | 370 ms |

### Req / Sec
| #Run | 1% | 2.5% | 50% | 97.5% | Avg | Stdev | Min |
| :--- | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
| 1 | 10615 | 10615 | 23615 | 29359 | 21946.4 | 6194.16 | 10608 |
| 2 | 12415 | 12415 | 29183 | 30527 | 25704.8 | 6797.97 | 12413 |
| 3 | 12039 | 12039 | 29567 | 30447 | 25402.4 | 6972.07 | 12035 |

### Bytes / Sec
| #Run | 1% | 2.5% | 50% | 97.5% | Avg | Stdev | Min |
| :--- | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
| 1 | 1.98 MB | 1.98 MB | 4.42 MB | 5.49 MB | 4.1 MB | 1.16 MB | 1.98 MB |
| 2 | 2.32 MB | 2.32 MB | 5.46 MB | 5.71 MB | 4.81 MB | 1.27 MB | 2.32 MB |
| 3 | 2.25 MB | 2.25 MB | 5.53 MB | 5.69 MB | 4.75 MB | 1.3 MB | 2.25 MB |